+++ kipper/core/src/compiler/lib/import/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Import folder including all importable kipper libraries that are provided per default to all Kipper programs.
+ * @author Luna Klatzer
+ * @copyright 2021-2022 Luna Klatzer
+ * @since 0.7.0
+ */
+