+++ kipper/core/src/compiler/target/index.ts
@@ -1,7 +1,0 @@
-/**
- * Targets for the Kipper language.
- * @since 0.5.0
- */
-
-export * as typescript from "./typescript";
-export * from "./compile-target";